% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptclust.R
\name{adapt.clust}
\alias{adapt.clust}
\title{Clustering by Binning}
\usage{
adapt.clust(data, nbins = 4, bin.defs = NULL, markers, dist = "eucl",
  growfact = 3, dist.thresh = 4, maxit = 100, minpts = 5)
}
\arguments{
\item{data}{data.frame or matrix containing the data to be clustered}

\item{nbins}{number of bins created at the start. If set to NULL and bin.mat is given, then starting nodes will be initialised using a perumuation matrix based on the positive and negative expression values given in bin.mat. Default is 4.}

\item{bin.defs}{2*length(markers) numeric matrix giving the expression values for positive and negative phenotypes for each marker. If set to NULL, the number of initial bins will be determined by nbins using the density distribution of the data. Default is NULL.}

\item{markers}{vector of marker names in data to be clustered}

\item{dist}{character vector indicating which distance metric to be used; one of "eucl" = euclidean, "jacc" = Jaccard.}

\item{growfact}{numeric indicates the number of new nodes created if average dist is above threshold}

\item{dist.thresh}{numeric minimum average distance threshold required to prevent a node splitting and forming new nodes.}

\item{maxit}{integer maximum allowed interations of the algorithm}

\item{minpts}{minimum node membership required for resampling the node def from the old node members.}
}
\description{
Cluster large datasets that cannot ordinarily be clustered by distance calculation because of size limitations. It works by creating bins into which observations are placed by nearest distance, currently euclidean distance is implemented.
}
\author{
Julian Spagnuolo
}
