% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binclust_it.R
\name{binclust.it}
\alias{binclust.it}
\title{Bin Clustering}
\usage{
binclust.it(expdata, markers, clust.col, noise.clust.id = "0",
  minpts = 4, alpha = 0.001, maxit = 5)
}
\arguments{
\item{markers}{Character vector, identifies vectors in data containing the marker expression data.}

\item{clust.col}{Character vector, column name identifying the vector of cluster ids.}

\item{noise.clust.id}{If a noise cluster is present in the data, this parameter will remove it from sub-clustering. Ie. DBScan identifies noise points as cluster "0". If NULL, this parameter is not used. Default is NULL.}

\item{minpts}{integer. Minimum number of cells a cluster must contain in order to be kept. Default is 4.}

\item{alpha}{numeric. alpha param for qchisq call to identify cluster outliers by mahalanobis distance (see qchisq)}

\item{maxit}{integer. Maximum number of iterations to run. Default is 5.}

\item{x}{data.frame. Contains marker expression and cluster id data.}

\item{bin.list}{list of matrices. Each member of the list is a matrix defining the sub-clusters within each original clusters that should be split. Output from binmat function.}
}
\description{
Bin Clustering
}
\author{
Julian Spagnuolo
}
