% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.select.R
\name{bw.select}
\alias{bw.select}
\title{Bandwidth Select}
\usage{
bw.select(x, scalest = "mad", level = 2L, kernel = "normal",
  canonical = FALSE, gridsize = 401L, range.x = range(x),
  truncate = TRUE)
}
\description{
Bandwidth Select
}
\author{
Julian Spagnuolo
}
