% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellNetwork.R
\name{cell.network}
\alias{cell.network}
\title{Cell Network}
\usage{
cell.network(data, markers, clusters, phenotypes, state.clusters,
  major.phenos, halos, distMethod = "euclidean", parallel = FALSE,
  cores = NULL)
}
\description{
Cell Network
}
\author{
Julian Spagnuolo
}
