% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustDiff.R
\name{clust.diff}
\alias{clust.diff}
\title{Cluster Diff}
\usage{
clust.diff(data, markers, cluster.id, test.method = "Chisq",
  p.adj = TRUE, padj.method = "BH")
}
\arguments{
\item{data}{data.frame of observations, column names must match names in markers}

\item{markers}{character vector, length must match number of columns in data}

\item{cluster.id}{character vector of length 1 matching the name of the column in data corresponding to the numeric cluster.id vector.}

\item{test.method}{Character vector of length 1. Type of test to use in anova step. Default is "Chisq", see \link{\code{help("anova")}} for details.}

\item{p.adj}{Logical. Whether to apply multiple testing correction or not. Default is TRUE.}

\item{padj.method}{Character vector of length 1. Method of multiple testing correction to apply. Default is "BH" for Benjamini-Hochberg/FDR method, see \code{\link{help("p.adjust")}} for other method.}
}
\description{
Cluster Diff
}
\author{
Julian Spagnuolo
}
