% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_split.R
\name{clust.split}
\alias{clust.split}
\title{Cluster Split}
\usage{
clust.split(x, markers, clusters)
}
\arguments{
\item{x}{data.frame or matrix of numeric expression data}

\item{markers}{character vector of column names in x to use in the cluster splitting}

\item{clusters}{character vector of cluster ids. length must equal nrow of x.}
}
\description{
Cluster Split
}
\author{
Julian Spagnuolo
}
