% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfind2.R
\name{cutfind2}
\alias{cutfind2}
\title{@title Cutfind
 @description finds summary stats for distributions by either mixture modelling or finding the mode and mad of unimodal distributions
 @author Julian Spagnuolo}
\usage{
cutfind2(x, markers, npeaks = NULL, DensityThreshold = NULL,
  gridsize = 14000, max.restarts = 1, epsilon = 1e-09,
  maxit = 8000, seed = 42, auto = TRUE, metric = "AIC",
  which.dev = 2)
}
\description{
@param x data.frame or named matrix containing raw data for which you wish to identify cutoff values for subsequent transformation.
 @param markers a vector containing the columns in x that you wish to parse through cutfind2
 @param npeaks a named vector of integers indicating the number of peaks in each distribution that you wish to identify. This can contain NA values. See details
 @param DensityThreshold Named numeric vector indicating the density above which peaks will be identified. Set this below the smallest peak in the density distribution of your data.
 @param gridsize gridsize used in the density estimation steps - this needn't be changed.
 @param epsilon accuracy of mixture modelling step (only used if npeaks for marker is > 1)
 @param maxit maximum iterations used by mixture modelling step (only used if npeaks for marker is > 1). If this step fails, it will restart (number of restarts set by max.restarts) and add 2000 iterations to maxit.
 @param max.restarts maximum restarts allowed in the mixture modelling steps (increases maxit by 2000 each restart). The function will use the parameters returned by the mixture modelling step to re-seed the algorithm.
 @param seed This is the system seed used by the mixture modelling step, just to increase the reproducibility of the function
 @param auto Logical - if TRUE the function will search for the optimal number of components (distributions) to fit in the mixture model by searching a range of +/- 2 peaks from the number of peaks identified.
 @param metric vector - the metric used to choose the optimum mixture model to fit to the data choose one of "BIC", "AIC", or "ICL"
 @param which.dev integer - the number of deviations from the peak at which to set the cutoff value. Default = 2
 @details
}
