% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoffs.R
\name{cutoffs}
\alias{cutoffs}
\title{cutoffs}
\usage{
cutoffs(cut.data, markers, component = NULL)
}
\arguments{
\item{cut.data}{output from cutfind function}

\item{markers}{vector of marker names}

\item{component}{named vector of the specific component from which the cutoff should be extracted. Names should be identical to markers}
}
\description{
cutoffs
}
\author{
Julian Spagnuolo
}
