% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan_opt.R
\name{dbscan.opt}
\alias{dbscan.opt}
\title{DBscan Parameter Scanning}
\usage{
dbscan.opt(data, eps.start, eps.end, step.size, minPts.start = NULL,
  minPts.end = NULL)
}
\arguments{
\item{data}{continuous numeric data to pass to dbscan}

\item{eps.start}{numeric, epsilon to begin scan from}

\item{eps.end}{numeric, epsilon to scan to}

\item{step.size}{numeric, rate at which to increase the epsilon parameter}

\item{minPts.start}{integer, minimum k nearest neighbours for a cluster, if NULL defaults to ncol(data)+1.}

\item{minPts.end}{integer, minimum k nearest neighbours to scan to. If defined, function will scan all epsilons between eps.start and eps.end for all minPts between minPts.start and minPts.end. Default is NULL.}
}
\value{
data.frame. eps = vector of epsilons scanned through; minPts = vector of minPts scanned through; n.clust = number of clusters found in each epsilon step for minPts; noise.pts = number of points considered to be noise at each step; log.noise.pts = log10 of the number of noise points (included for plotting).
}
\description{
DBscan Parameter Scanning
}
\author{
Julian Spagnuolo
}
