% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichTest.R
\name{enrichTest}
\alias{enrichTest}
\title{Enrichment Testing}
\usage{
enrichTest(x, clust.col, noise.clust.id = NULL, cat.col,
  equal.props = FALSE, alternative = "greater")
}
\arguments{
\item{x}{data.frame/matrix. Data to be tested against, should contain column of cluster IDs and categorical/annotation column.}

\item{clust.col}{character vector. Column name in x containing the cluster ID's.}

\item{noise.clust.id}{character. ID of noise clusters if present (i.e DBscan ID's noise as "0"). If NULL this will not be used. Default is NULL.}

\item{cat.col}{character vector. Column name in x with the categorical (factor) data to be tested for enrichment.}

\item{equal.props}{logical/boolean. If TRUE binomial test will use an equal hypothesised probability of success in the binomial test, if FALSE binomial test uses the global proportion (including the noise points if present). Default is FALSE. Note: The hypergeometric test will always use the global proportion.}

\item{alternative}{character. Indicates alternative hypothesis of binomial test. Must be one of "two.sided", "greater", or "less". You can specify just the initial letter.Default is "greater"}
}
\description{
Enrichment Testing
}
\details{
For the binomial test,
}
\author{
Julian Spagnuolo
}
