% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucl.R
\name{eucl}
\alias{eucl}
\title{Euclidean Distance of a Point to a Vector}
\usage{
eucl(x, y)
}
\arguments{
\item{x}{a numeric vector representing the origin from which to calculate the distance to y}

\item{y}{a data.frame or matrix representing the points to which you want to calculate the distance from x. ncol(y) must equal length(x).}
}
\description{
Euclidean Distance of a Point to a Vector
}
\author{
Julian Spagnuolo
}
