% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facsnorm.R
\name{facsnorm}
\alias{facsnorm}
\title{Facsnorm}
\usage{
facsnorm(x, cutoffs, asinCofac = 25, method = c("arcsin", "loglin",
  "logit"))
}
\arguments{
\item{x}{data.frame or matrix of raw data}

\item{cutoffs}{vector of cutoff values. values MUST be in the order of the markers in x that they are to be applied to}

\item{asinCofac}{spreading parameter for arcsin transformation. Only needs to be set if method == arcsin. Default == 25.}

\item{method}{Type of transformation to be applied, either arcsin ("arcsin"), log-linear ("loglin") or logit ("logit") transformation}
}
\description{
Facsnorm
}
\author{
Julian Spagnuolo, Tobias Rutishauser
}
