% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_cov.R
\name{med.cov}
\alias{med.cov}
\title{Median Covariance}
\usage{
med.cov(expdata, markers, use.median = TRUE)
}
\arguments{
\item{expdata}{data.frame. Contains expression data}

\item{markers}{Character vector, identifies vectors in data containing the marker expression data.}

\item{use.median}{Logical. If TRUE, the median of columns in expdata is used to calculate the covariance. If FALSE, the mean is used (identical to stats::cov). Default is TRUE.}
}
\description{
Median Covariance
}
\author{
Julian Spagnuolo
}
