% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeSize.R
\name{nodeSize}
\alias{nodeSize}
\title{Scaled Node Sizes for growing SOM}
\usage{
nodeSize(somtree, reset = FALSE, maxNodeSize = 0.1)
}
\arguments{
\item{somtree}{results from calling som.tree}

\item{reset}{Logical. If TRUE, node size is reset to maxNodeSize, default = FALSE}

\item{maxNodeSize}{Maximum size for any node. Default = 15}
}
\description{
calcs size of nodes for som tree
}
\author{
Julian Spagnuolo
}
