% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\title{Posterior Probability}
\usage{
posterior(model.params, data)
}
\arguments{
\item{model.params}{list, output of phenosampler}

\item{data}{numeric matrix or data.frame. ncols must equal length(names(model.params))}
}
\description{
Posterior Probability
}
\author{
Julian Spagnuolo
}
