% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saneMarkers.R
\name{saneMarkers}
\alias{saneMarkers}
\title{Sanitizing Marker Names}
\usage{
saneMarkers(markers)
}
\arguments{
\item{markers}{Character vector, taken from first row of each FACS data file imported into FACkit GUI}
}
\description{
Sanitizing Marker Names
}
\author{
Julian Spagnuolo
}
