% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somtree.R
\name{som.tree}
\alias{som.tree}
\title{Build Growing Self-organising Map and Construct a Tree}
\usage{
som.tree(data, markers, shape = c("rect", "hex"), maxit = 500,
  cores = 1, alpha = 0.9, beta = 0.5, spread = 0.9,
  alg = c("kohonen"), dim = c(NULL, NULL))
}
\arguments{
\item{data}{data.frame or matrix containing expression data}

\item{markers}{vector of markers/columns in data to build the SOM with}

\item{shape}{vector determining the shape of SOM to build, either rectangular, "rect", or hexagonal, "hex".}

\item{maxit}{number of iterations the SOM building function will use, default = 500.}

\item{cores}{number of cores to using in distance computation. Default = 1, set higher if you have more than 1 cpu to use.}

\item{alpha}{discount factor for the learning rate during the growing phase of the training. Values should be between 0 and 1. Default = 0.9}

\item{beta}{propagation rate. Determines the rate at which the error of a node, that cannot grow any nodes, is passed on to its neighbours. Suggested values may range between 0 and 1. Default = 0.5}

\item{spread}{numeric value between 0 and 1, controls the rate at which new nodes are created in the growing SOM algorithm, lower values decrease the rate, higher ones increase it. Default is 0.9}

\item{alg}{character vector of length 1. Determines which SOM algorithm to use, choices include either Kohonen SOM, "kohonen" or Growing SOM, "grow" - GrowingSOM is not implemented - awaiting a bug fix and rerelease.}

\item{dim}{integer vector of length 2. Determines dimensions of the SOM in the x and y dimension, respectively. These parameters must be set for the kohonen algorithm. If using the GrowSOM algorithm, defining dim will force the growing SOM to use a gridsize of dimension dim[1]*dim[2]. Default is c(NULL,NULL).}
}
\description{
Build Growing Self-organising Map and Construct a Tree
}
\author{
Julian Spagnuolo
}
