% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_merge.R
\name{splitmerge}
\alias{splitmerge}
\title{Split Merge}
\usage{
splitmerge(x, markers, clust.col, bin.list, noise.clust.id = NULL)
}
\arguments{
\item{x}{data.frame. Contains marker expression and cluster id data.}

\item{markers}{Character vector, identifies vectors in data containing the marker expression data.}

\item{clust.col}{Character vector, column name identifying the vector of cluster ids.}

\item{bin.list}{list of matrices. Each member of the list is a matrix defining the sub-clusters within each original clusters that should be split. Output from binmat function.}

\item{noise.clust.id}{If a noise cluster is present in the data, this parameter will remove it from sub-clustering. Ie. DBScan identifies noise points as cluster "0". If NULL, this parameter is not used. Default is NULL.}
}
\description{
Split Merge
}
\author{
Julian Spagnuolo
}
