% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholder.R
\name{thresholder}
\alias{thresholder}
\title{Threshold Phenotyping}
\usage{
thresholder(data, markers, thresholds)
}
\arguments{
\item{data}{numeric matrix/data.frame. Contains FACS marker expression data}

\item{markers}{character vector. Names of markers in data to phenotype using thresholds}

\item{thresholds}{named numeric vector. Expression cut points on which to call positive (1) or negative (0) phenotypes.}
}
\description{
Threshold Phenotyping
}
\author{
Julian Spagnuolo
}
