% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsnehm.R
\name{tsnehm}
\alias{tsnehm}
\title{Compute t-SNE Heatmap}
\usage{
tsnehm(data, markers, tsne_embedding, cell_labels, breaks = 100,
  plot = TRUE)
}
\arguments{
\item{markers}{markers of interest}

\item{tsne_embedding}{1D t-SNE embedding}

\item{cell_labels}{Labels for each cell. Typically these are the cluster assignments for each cell, as obtained by 
dbscan on the 2D t-SNE. This way, the columns of the heatmap will have a color assigned to each of them,
and they can be mapped to a corresponding location on the 2D t-SNE}

\item{breaks}{Number of bins}

\item{plot}{Logical. Whether to plot the heatmap (TRUE). If false, a list containing a matrix of the heatmap data and the column annotations (group_label) will be returned. Default is TRUE}

\item{expression_matrix}{FACS data. Columns should be markers, rows are cells.}
}
\description{
Experimental version of t-SNE heatmaps codes
}
